Fix for Linux (unfix a warning on HP/UX 11 CC)
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/104



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@7253 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/test/intltest/itrbnf.cpp
@@ -338,7 +338,7 @@ void IntlTestRBNF::TestLLongConstructors()
     // copy constructor
     LLAssert(llong(llong(0, 1)).asDouble() == 1);
     LLAssert(llong(llong(1, 0)).asDouble() == llong::kD32);
-    LLAssert(llong(llong((uint32_t)-1, (uint32_t)-1)).asDouble() == -1);
+    LLAssert(llong(llong(-1, (uint32_t)-1)).asDouble() == -1);
 
     // asInt - test unsigned to signed narrowing conversion
     LLAssert(llong((uint32_t)-1).asInt() == (int32_t)0x7fffffff);
